requirements.txt
discord.py>=2.4,<3
python-dotenv>=1.0,<2
deep-translator>=1.11,<2
aiohttp>=3.9,<4
